backend/requirements.txt
bcrypt==4.3.0
email-validator==2.3.0
fastapi==0.116.1
PyJWT==2.10.1
pydantic-settings==2.10.1
SQLAlchemy==2.0.43
uvicorn[standard]==0.35.0

